+++ CHANGELOG.md
@@ -14,3 +14,13 @@
 - Added custom annotators for JavaScript and TypeScript to match VSCode Style
 - Added `Settings` -> `Editor` -> `Color Scheme` - > `JavaScript (Extra)`
   > Note - This will be used by both JavaScript and TypeScript
+
+## [1.1.0]
+### Added
+- Added extensive syntax highlighting for Java. (Provides same color)
+- Added extensive syntax highlighting for Dart/Flutter.
+### Fixed
+- (Bug) - Fixed PSIElement font color & Keywords font color to be same as VSCode.
+### Changed
+- start highlight onTyping
+- Documentation on tips for best viewing experience.